feat(refactor): move test classes package
Move each test classes package to model package.
